feat: 북마크조회 콤포넌트 생성, 북마크버튼 BoardHeader로 이동

+++ src/components/Caption/CommentItem.jsx
@@ -9,7 +9,7 @@ import {
   CommentBoxItemIconButton,
 } from "./styles/CommentItem.styled";
 
-import { faPenToSquare, faTrash } from '@fortawesome/free-solid-svg-icons';
+import { faPenToSquare, faTrash } from "@fortawesome/free-solid-svg-icons";
 import { CommentBoxButton, CommentBoxTextArea } from "./styles/Comment.styled";
 
 const CommentItem = ({ userId, item }) => {
@@ -17,29 +17,40 @@ const CommentItem = ({ userId, item }) => {
   const [editedContent, setEditedContent] = useState();
 
   return (
-    <CommentBoxItemWrapper id={item.id} key={item.id}user={item.userId} board={item.boardId}>
+    <CommentBoxItemWrapper
+      id={item.id}
+      key={item.id}
+      user={item.userId}
+      board={item.boardId}
+    >
       <CommentBoxItemTitle>
-          <CommentBoxItemName>{item.username}</CommentBoxItemName>
-          <CommentBoxItemTime>(작성 {item.created_at} 수정 {item.created_at})</CommentBoxItemTime>
-        <CommentBoxItemIconButton icon={faPenToSquare} 
-            onClick={() => {
-              setIsEditing(!isEditing);
-              setEditedContent(item.content);
-            }}
-          >
-        </CommentBoxItemIconButton>
-        <CommentBoxItemIconButton icon={faTrash} 
+        <CommentBoxItemName>{item.username}</CommentBoxItemName>
+        <CommentBoxItemTime>
+          (작성 {item.created_at} 수정 {item.created_at})
+        </CommentBoxItemTime>
+        <CommentBoxItemIconButton
+          icon={faPenToSquare}
+          onClick={() => {
+            setIsEditing(!isEditing);
+            setEditedContent(item.content);
+          }}
+        ></CommentBoxItemIconButton>
+        <CommentBoxItemIconButton
+          icon={faTrash}
           type="submit"
           onClick={(e) => {
-            const li = e.currentTarget.closest("li")
-            const body = {
-                id: li.id,
-                user_id: userId
+            if (userId != item.userId) {
+              alert("작성자가 다릅니다.");
+              return;
             }
+            const li = e.currentTarget.closest("li");
+            const body = {
+              id: li.id,
+              user_id: userId,
+            };
             deleteComment(JSON.stringify(body));
           }}
-        >
-        </CommentBoxItemIconButton>
+        ></CommentBoxItemIconButton>
       </CommentBoxItemTitle>
       <CommentBoxItemContent>{item.content}</CommentBoxItemContent>
       {isEditing ? (
@@ -53,6 +64,10 @@ const CommentItem = ({ userId, item }) => {
           <br />
           <CommentBoxButton
             onClick={() => {
+              if (userId != item.userId) {
+                alert("작성자가 다릅니다.");
+                return;
+              }
               const body = {
                 id: item.id,
                 user_id: userId,
@@ -68,9 +83,7 @@ const CommentItem = ({ userId, item }) => {
           </CommentBoxButton>
         </>
       ) : (
-        <>
-          
-        </>
+        <></>
       )}
     </CommentBoxItemWrapper>
   );
